requirements.txt
requests
python-dotenv
python-telegram-bot
pandas
beautifulsoup4
flask
